refactor: 모달 폴더 구조 변경
- lib 폴더 내부에 있는 폴더들 외부로 분리
