+++ components/src/App.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import "./App.css";
-import Modal from "./lib/components/Modal/Modal";
+import Modal from "./components/Modal/Modal";
 
 function App() {
   const [isOpen, setIsOpen] = useState(false);
