+++ components/src/lib/index.ts
@@ -1,1 +1,1 @@
-export { default as Modal } from "./components/Modal/Modal";
+export { default as Modal } from "../components/Modal/Modal";
